Cleaned up code, revised documentation.
Resolved issues that were causing the following test to fail:

```
test_mock_method_arguments_returns_an_array_of_mock_arguments_of_the_correct_type_for_the_specified_method_of_the_class_or_object_instance_reflected_by_the_reflection_returned_by_the_mock_class_instances_reflection_method

```

It is not possible to mock arguments for a method that expects
an implementation of a interface or abstract class since it
is not possible to instantiate a interface or abstract class.

The test now expects a `RuntimeException` to be thrown if
`MockClassInstance->mockMethodArguments` is called to mock
method arguments for a method that defines one or more
parameters that expect an implementation of an interface
or an abstract class.